[Behat][Core] Moved the EmailChecker to Core component
