+++ app/config/config_test.yml
@@ -10,11 +10,6 @@ parameters:
 framework:
     test: ~
     default_locale: "en"
-    profiler:
-        matcher:
-            path: ^/shop/forgotten-password|^/shop/verify|^/shop/register
-        only_exceptions: false
-        collect: true
     session:
         storage_id: session.storage.mock_file
