+++ src/Sylius/Component/Core/Test/Services/EmailCheckerInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Behat\Service\Accessor;
+namespace Sylius\Component\Core\Test\Services;
 
 /**
  * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
